requirements.txt
asgiref==3.11.1
brotli==1.2.0
click==8.3.1
colorama==0.4.6
dj-database-url==3.1.2
Django==6.0.2
djangorestframework==3.16.1
gunicorn==25.1.0
h11==0.16.0
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
tzdata==2025.3
uvicorn==0.41.0
whitenoise==6.11.0
